requirements.txt
absl-py==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
ansi2html==1.6.0
anyio==3.3.4
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==2.1.3
argcomplete==2.0.0
arviz==0.11.2
asgiref==3.4.1
async-timeout==4.0.2
atari-py==0.2.9
atomicwrites==1.1.5
attrs==19.3.0
backcall==0.2.0
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==1.2.2
blinker==1.4
blis==0.7.4
Box2D==2.3.10
Box2D-kengz==2.3.3
Brlapi==0.7.0
Brotli==1.0.9
buildozer==0.40.dev0
cachetools==4.2.2
catalogue==2.0.4
certifi==2019.11.28
cffi==1.15.0
cftime==1.5.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
cytoolz==0.11.2
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
dbus-python==1.2.16
debugpy==1.4.1
decorator==4.4.2
defer==1.0.6
dill==0.3.4
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-static==4.2.1
django-core==1.4.1
django-cors-headers==3.11.0
django-forms-bootstrap==3.1.0
django-storages==1.12.3
django-versatileimagefield==2.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
dm-tree==0.1.6
dnspython==2.1.0
duplicity==0.8.12.0
EasyProcess==0.3
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.0.1
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastapi==0.70.0
fasteners==0.17.3
fastprogress==1.0.0
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Compress==1.10.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
gin-config==0.4.0
glfw==2.5.0
greenlet==1.1.0
gym==0.18.3
h11==0.12.0
hexbytes==0.2.2
html5lib==1.0.1
httplib2==0.14.0
httptools==0.2.0
idna==2.8
imageio==2.4.0
importlib-metadata==1.5.0
ipfshttpclient==0.8.0a2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jdcal==1.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-dash==0.4.0
keyring==18.0.1
Kivy==1.10.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.1.2
matplotlib-inline==0.1.2
mediapipe==0.8.4.2
monotonic==1.5
more-itertools==4.2.0
mujoco-py==2.1.2.14
multiaddr==0.0.9
multidict==5.2.0
murmurhash==1.0.5
netaddr==0.8.0
netCDF4==1.5.7
netifaces==0.10.4
nltk==3.6.2
numexpr==2.7.1
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.54
openpyxl==3.0.3
packaging==20.3
pandas==0.25.3
paramiko==2.6.0
parsimonious==0.8.1
parso==0.8.2
pathy==0.5.2
patsy==0.5.1
pep517==0.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipx==1.0.0
plotly==5.1.0
pluggy==0.13.0
preshed==3.0.5
prompt-toolkit==3.0.19
protobuf==3.17.1
py==1.8.1
py-solc-x==1.1.1
py4j==0.10.9.3
pyarrow==7.0.0
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.12.0
pycups==1.9.73
pydantic==1.7.4
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymc3==3.11.2
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.2.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.6
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-magic==0.4.25
pytoml==0.1.21
pytz==2019.3
PyVirtualDisplay==2.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
regex==2021.4.4
reportlab==3.6.6
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rlp==2.0.1
scipy==1.3.3
SecretStorage==2.3.1
selenium==3.141.0
semantic-version==2.8.5
semver==2.13.0
sh==1.14.2
simplejson==3.16.0
six==1.14.0
smart-open==3.0.0
sniffio==1.2.0
soupsieve==1.9.5
spacy==3.1.1
spacy-legacy==3.0.8
SQLAlchemy==1.4.16
sqlparse==0.4.1
srsly==2.4.1
stable-baselines==2.10.2
starlette==0.16.0
systemd-python==234
tables==3.6.1
tenacity==8.0.1
tensorflow-probability==0.12.2
textblob==0.15.3
tf-agents==0.8.0
Theano-PyMC==1.1.2
thinc==8.0.8
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
userpath==1.8.0
uvicorn==0.15.0
uvloop==0.16.0
variety==0.8.3
varint==1.0.2
virtualenv==20.4.7
wadllib==1.3.3
wasabi==0.8.2
watchgod==0.7
wcwidth==0.1.8
web3==5.25.0
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.1
wordcloud==1.8.1
wrapt==1.12.1
WTForms==2.3.3
xarray==0.19.0
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
yarl==1.7.2
zipp==1.0.0
